Feat prerequisite (#39)
* attack

* ACTot

* aC

* att

* attack

* bab

* feats

* enum

* show

* addFeat

* skill

* hp

* vitaError

* move

* sizeAr

* virgola

* sizeAc

* dto

* errorFeats

* attSize

* szSk

* grapple

* grapple

* specialAttack

* checkFeat

* featsList

* dupli

* featcheck

* featcheck

* feat

* switch

* feat

* rage49

* dupl

* dupli_DTO_method

* prerequisiteDTO

* sawintThrow_check

* sawintThrow_check_1

* Prerequisite_checkingi

* Prerequisite_checkingiV2

* empty_prer

* prerequisite_json

* characterF_dupli_classVSbuy

* add_featFromClassVSbuy

* checking_pre

* checking_preV2

* checking_preV3

+++ api/src/main/java/pl/kolendateam/dadcard/skills/dto/SkillsDTO.java
@@ -5,7 +5,7 @@ public class SkillsDTO {
     public int idSkill;
     public String nameSkill;
     public boolean classSkill;
-    public int skillRank;
+    public double skillRank;
     public int skillAbility;
     public int skillBonus;
